feat/28: move retrofitinstance file
