+++ app/src/main/java/com/devdi/mapmories/RetrofitInstance.kt
@@ -1,10 +1,10 @@
-package com.devdi.mapmories.login
+package com.devdi.mapmories
 
+import com.devdi.mapmories.login.LoginApiService
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import java.nio.charset.StandardCharsets
 
 object RetrofitInstance {
     private const val BASE_URL = "http://43.201.246.74:8080/"
